+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+- [PR-27](https://github.com/itk-dev/ai-screening/pull/27)
+  Added project node template
 - [PR-26](https://github.com/itk-dev/ai-screening/pull/26)
   OpenID Connect changes
 - [PR-25](https://github.com/itk-dev/ai-screening/pull/25)